Backend/requirements.txt
Flask
Flask-PyMongo
Flask-Bcrypt
Flask-JWT-Extended
Flask-CORS
python-dotenv
scikit-learn
pandas
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform
googlemaps
requests
gunicorn